pyproject.toml
[project]
name = "isca"
version = "0.1.0"
description = "Integrated Symbolic Cognition Architecture"
authors = [
    {name = "jacob",email = "jacobfv123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "torch",
    "transformers",
    "tqdm",
    "numpy",
    "pyyaml",
    "networkx",
    "agisdk",
    "langchain",
    "langchain-openai",
    "langchain-anthropic", 
    "langchain-google-genai",
    "langchain-huggingface",
    "accelerate",
    "safetensors",
    "datasets",
    "scipy",
    "matplotlib",
    "seaborn",
    "scikit-learn",
]

[tool.poetry]
packages = [{include = "isca", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
